+++ src/interfaces/http/http_client.rs
@@ -1,7 +1,8 @@
 use crate::notifications::Notification;
 use crate::LIB_LOG_TARGET;
-use log::{debug, error, warn};
+use log::{debug, error, trace, warn};
 use reqwest::Client;
+use tokio::sync::broadcast::error::TryRecvError;
 use tokio::sync::{broadcast, watch};
 
 pub(super) async fn start_sending(
@@ -9,33 +10,42 @@ pub(super) async fn start_sending(
     shutdown: watch::Receiver<bool>,
     url: &str,
 ) {
-    let mut shutdown_rx = shutdown.clone();
     let mut rx = interface_rx.resubscribe();
     let client = Client::new();
 
     loop {
-        tokio::select! {
-            received = rx.recv() => {
-                match received {
-                    Ok(message) => {
-                        let response = client.post(url)
-                        .body(message.to_json().unwrap_or_default())
-                        .send().await;
-                        match response {
-                            Ok(ok) => debug!(target: LIB_LOG_TARGET,"HTTP Client Response - status: {} url: {}", ok.status(), ok.url()),
-                            Err(error) => warn!(target: LIB_LOG_TARGET, "HTTP Client Response Error: {}", error ),
-                        }
-                    },
-                    Err(error) => {
-                        error!(target: LIB_LOG_TARGET, "Broadcast Receiver Error: {}", error);
-                        break;
-                    },
+        if *shutdown.borrow() {
+            trace!(target: LIB_LOG_TARGET, "HTTP Client start_sending received shutdown");
+            break;
+        }
+
+        match rx.try_recv() {
+            Ok(message) => {
+                debug!(target: LIB_LOG_TARGET, "HTTP Client received message");
+                let response = client.post(url).body(message.to_json().unwrap_or_default()).send().await;
+                match response {
+                    Ok(ok) => debug!(
+                        target: LIB_LOG_TARGET,
+                        "HTTP Client Response - status: {} url: {}",
+                        ok.status(),
+                        ok.url()
+                    ),
+                    Err(error) => warn!(target: LIB_LOG_TARGET, "HTTP Client Response Error: {}", error),
                 }
             }
 
-            _ = shutdown_rx.changed() => {
-                break;
-            }
+            Err(error) => match error {
+                TryRecvError::Empty => {
+                    trace!(target: LIB_LOG_TARGET, "HTTP Client Broadcast Receiver: {}", error);
+                }
+                TryRecvError::Closed => {
+                    error!(target: LIB_LOG_TARGET, "HTTP Client Broadcast Receiver Error: {}", error);
+                    break;
+                }
+                TryRecvError::Lagged(skipped) => {
+                    warn!(target: LIB_LOG_TARGET, "Broadcast Receiver: {} messages skipped: {}", error, skipped);
+                }
+            },
         }
     }
 }